+++ apps/rosen-service2/src/app.ts
@@ -3,6 +3,7 @@ import { ServiceManager } from '@rosen-bridge/service-manager';
 
 import dataSource from './data-source';
 import { DBService } from './services/db';
+import { ErgoScannerService } from './services/ergoScanner';
 
 const logger = CallbackLoggerFactory.getInstance().getLogger(import.meta.url);
 
@@ -19,8 +20,13 @@ const startApp = async () => {
   serviceManager.register(DBService.getInstance());
   logger.debug('Database service registered to the service manager');
 
+  logger.debug('Initializing scanner service');
+  await ErgoScannerService.init(DBService.getInstance());
+  serviceManager.register(ErgoScannerService.getInstance());
+  logger.debug('Scanner service registered to the service manager');
+
   logger.debug('Starting service manager...');
-  await serviceManager.start(DBService.getInstance().getName());
+  await serviceManager.start(ErgoScannerService.getInstance().getName());
 };
 
 export default startApp;